loader: Add an ISO9660 "partition table" type to loader.
Update from FreeBSD: https://svnweb.freebsd.org/base?view=revision&revision=332085

When booted via isoboot(8) loader will be handed a disk that simply contains
an ISO9660 image. Currently this confuses it greatly. Teach it how to spot
that it's in this situation and that ISO9660 has one "partition" covering
the whole disk.