loader: Add isoboot(8) for booting BIOS systems from HDDs containing ISO images.
FreeBSD update: https://svnweb.freebsd.org/base?view=revision&revision=332084

This is part of a project for adding the ability to create hybrid CD/USB boot
images. In the BIOS case when booting from something that isn't a CD we need
some extra boot code to actually find our next stage (loader) within an
ISO9660 filesystem. This code will reside in a GPT partition (similar to
gptboot(8) from which it is derived) and looks for /boot/loader in an
ISO9660 filesystem on the image.